+++ backend/notification-service/src/main/java/com/github/spjavaind300/notificationservice/service/imp/ProfileServiceImp.java
@@ -16,7 +16,7 @@ public class ProfileServiceImp implements ProfileService {
     private final ProfileServiceClient client;
 
     @Override
-    public List<UserDto> getProfiles(List<Integer> userIds) {
+    public List<UserDto> getProfiles(List<Long> userIds) {
         return new ArrayList<>(client.getProfiles(userIds));
     }
 }
